+++ packages/map-template/src/atoms/kioskOriginLocationIdState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const kioskOriginLocationIdState = atom({
+    key: 'kioskOriginLocationId',
+    default: undefined
+});
+
+export default kioskOriginLocationIdState;