+++ packages/map-template/src/components/Search/Search.jsx
@@ -8,7 +8,6 @@ import { usePreventSwipe } from '../../hooks/usePreventSwipe';
 import ListItemLocation from '../WebComponentWrappers/ListItemLocation/ListItemLocation';
 import SearchField from '../WebComponentWrappers/Search/Search';
 import filteredLocationsState from '../../atoms/filteredLocationsState';
-import primaryColorState from '../../atoms/primaryColorState';
 import mapsIndoorsInstanceState from '../../atoms/mapsIndoorsInstanceState';
 import currentLocationState from '../../atoms/currentLocationState';
 import isLocationClickedState from '../../atoms/isLocationClickedState';
@@ -19,9 +18,14 @@ import languageState from '../../atoms/languageState';
 import { useTranslation } from 'react-i18next';
 import kioskLocationState from '../../atoms/kioskLocationState';
 import getDesktopPaddingBottom from '../../helpers/GetDesktopPaddingBottom';
+import { createPortal } from 'react-dom';
 import useKeyboardState from '../../atoms/useKeyboardState';
 import Keyboard from '../WebComponentWrappers/Keyboard/Keyboard';
 import searchInputState from '../../atoms/searchInputState';
+import searchResultsState from '../../atoms/searchResultsState';
+import selectedCategoryState from '../../atoms/selectedCategoryState';
+import Categories from './Categories/Categories';
+import { useIsKioskContext } from "../../hooks/useIsKioskContext";
 import { ReactComponent as Legend } from '../../assets/legend.svg';
 import showLegendDialogState from '../../atoms/showLegendDialogState';
 
@@ -30,15 +34,17 @@ import showLegendDialogState from '../../atoms/showLegendDialogState';
  *
  * @param {Object} props
  * @param {[[string, number]]} props.categories - All the unique categories that users can filter through.
- * @param {function} props.onSetSize - Callback that is fired when the search field takes focus.
+ * @param {function} props.onSetSize - Callback that is fired when the search field takes focus and when categories are clicked.
+ * @param {boolean} props.isOpen - Boolean that describes if the search page is open or not.
  *
  * @returns
  */
-function Search({ onSetSize }) {
+function Search({ onSetSize, isOpen }) {
 
     const { t } = useTranslation();
 
     const searchRef = useRef();
+    const scrollButtonsRef = useRef();
 
     /** Referencing the search field */
     const searchFieldRef = useRef();
@@ -47,22 +53,16 @@ function Search({ onSetSize }) {
     const keyboardRef = useRef();
 
     const [searchDisabled, setSearchDisabled] = useState(true);
-    const [searchResults, setSearchResults] = useState([]);
+    const [searchResults, setSearchResults] = useRecoilState(searchResultsState);
     const categories = useRecoilValue(categoriesState);
     const useKeyboard = useRecoilValue(useKeyboardState);
 
     /** Indicate if search results have been found */
     const [showNotFoundMessage, setShowNotFoundMessage] = useState(false);
 
-    /** Referencing the categories results container DOM element */
-    const categoriesListRef = useRef();
-
-    /** Determines which category has been selected */
-    const [selectedCategory, setSelectedCategory] = useState();
+    const [selectedCategory, setSelectedCategory] = useRecoilState(selectedCategoryState);
 
     const scrollableContentSwipePrevent = usePreventSwipe();
-
-    const primaryColor = useRecoilValue(primaryColorState);
 
     const [hoveredLocation, setHoveredLocation] = useState();
 
@@ -86,6 +86,8 @@ function Search({ onSetSize }) {
 
     const searchInput = useRecoilValue(searchInputState);
 
+    const isKioskContext = useIsKioskContext();
+
     const [, setShowLegendDialog] = useRecoilState(showLegendDialogState);
 
     /**
@@ -97,36 +99,6 @@ function Search({ onSetSize }) {
         window.mapsindoors.services.LocationsService.getLocations({
             categories: category,
         }).then(onResults);
-    }
-
-    /**
-     * Handles the click events on the categories list.
-     *
-     * @param {string} category
-     */
-    function categoryClicked(category) {
-        setSelectedCategory(category);
-        setSize(snapPoints.MAX);
-
-        if (selectedCategory === category) {
-            // If the clicked category is the same as currently selected, "deselect" it.
-            setSearchResults([]);
-            setSelectedCategory(null);
-
-            // Pass an empty array to the filtered locations in order to reset the locations.
-            setFilteredLocations([]);
-
-            // Check if the search field has a value and trigger the search again.
-            if (searchFieldRef.current.getValue()) {
-                searchFieldRef.current.triggerSearch();
-            }
-        } else if (searchFieldRef.current.getValue()) {
-            // If the search field has a value, trigger a research based on the new category.
-            searchFieldRef.current.triggerSearch();
-        } else {
-            // If the search field is empty, show all locations with that category.
-            getFilteredLocations(category);
-        }
     }
 
     /**
@@ -170,7 +142,7 @@ function Search({ onSetSize }) {
     }
 
     /**
-     * When search field is clicked, maximize the sheet size and set focus on the from field, 
+     * When search field is clicked, maximize the sheet size and set focus on the from field,
      * and if the useKeyboard prop is present, show the onscreen keyboard.
      * But wait for any bottom sheet transition to end before doing that to avoid content jumping when virtual keyboard appears.
      */
@@ -220,7 +192,9 @@ function Search({ onSetSize }) {
             mapsIndoorsInstance.setFloor(locationFloor);
         }
 
-        fitBoundsLocation(location, mapsIndoorsInstance, getBottomPadding(), getLeftPadding());
+        Promise.all([getBottomPadding(), getLeftPadding()]).then(([bottomPadding, leftPadding]) => {
+            fitBoundsLocation(location, mapsIndoorsInstance, bottomPadding, leftPadding);
+        });
     }
 
     /**
@@ -228,15 +202,17 @@ function Search({ onSetSize }) {
      * Calculate all cases depending on the kioskLocation id prop as well.
      */
     function getBottomPadding() {
-        if (isDesktop) {
-            if (kioskLocation) {
-                return getDesktopPaddingBottom();
+        return new Promise((resolve) => {
+            if (isDesktop) {
+                if (kioskLocation) {
+                    getDesktopPaddingBottom().then(padding => resolve(padding));
+                } else {
+                    resolve(0);
+                }
             } else {
-                return 0;
+                resolve(200);
             }
-        } else {
-            return 200;
-        }
+        });
     }
 
     /**
@@ -244,14 +220,36 @@ function Search({ onSetSize }) {
      * Calculate all cases depending on the kioskLocation id prop as well.
      */
     function getLeftPadding() {
-        if (isDesktop) {
-            if (kioskLocation) {
-                return 0;
+        return new Promise((resolve) => {
+            if (isDesktop) {
+                if (kioskLocation) {
+                    resolve(0);
+                } else {
+                    getDesktopPaddingLeft().then(padding => resolve(padding));
+                }
             } else {
-                return getDesktopPaddingLeft();
+                resolve(0);
             }
+        });
+    }
+
+    /**
+     * Calculate the CSS for the container based on context.
+     */
+    function calculateContainerStyle() {
+        if (searchResults.length > 0) {
+            let maxHeight;
+            if (isDesktop) {
+                // On desktop-sized viewports, we want the container to have a max height of 60% of the Map Template.
+                maxHeight = document.querySelector('.mapsindoors-map').clientHeight * 0.6 + 'px';
+            } else {
+                // On mobile-sized viewports, take up all available space if needed.
+                maxHeight = '100%';
+            }
+
+            return { display: 'flex', flexDirection: 'column', maxHeight, overflow: 'hidden' };
         } else {
-            return 0;
+            return { minHeight: categories.length > 0 ? '136px' : '80px' };
         }
     }
 
@@ -288,12 +286,22 @@ function Search({ onSetSize }) {
     });
 
     /*
-     * When useKeyboard parameter is present, add click event listener which determines when the keyboard should be shown or not. 
+     * Setup scroll buttons to scroll in search results list when in kiosk mode.
+     */
+    useEffect(() => {
+        if (isOpen && isKioskContext && searchResults.length > 0) {
+            const searchResultsElement = document.querySelector('.mapsindoors-map .search__results');
+            scrollButtonsRef.current.scrollContainerElementRef = searchResultsElement;
+        }
+    }, [searchResults, isOpen]);
+
+    /*
+     * When useKeyboard parameter is present, add click event listener which determines when the keyboard should be shown or not.
      */
     useEffect(() => {
         if (useKeyboard) {
             const onClick = (event) => {
-                // Use the closest() method to check if the element that has been clicked traverses the element and its parents 
+                // Use the closest() method to check if the element that has been clicked traverses the element and its parents
                 // until it finds a node that matches the 'mi-keyboard' selector.
                 // If the user clicks on the keyboard or the search fields, the keyboard should stay visible.
                 if (event.target.closest('mi-keyboard') ||
@@ -312,12 +320,19 @@ function Search({ onSetSize }) {
         }
     }, [useKeyboard]);
 
+
     return (
         <div className="search"
             ref={searchRef}
-            style={{ minHeight: categories.length > 0 ? '136px' : '80px' }}>
+            style={calculateContainerStyle()}>
+
+            { /* Search info which includes legend button if in a Kiosk context. */}
+
             <div className='search__info'>
                 <button className='search__legend' onClick={() => setShowLegendDialog(true)}><Legend /></button>
+
+                { /* Search field that allows users to search for locations (MapsIndoors Locations and external) */}
+
                 <SearchField
                     ref={searchFieldRef}
                     mapsindoors={true}
@@ -329,35 +344,56 @@ function Search({ onSetSize }) {
                     disabled={searchDisabled} // Disabled initially to prevent content jumping when clicking and changing sheet size.
                 />
             </div>
-            <div className="search__scrollable prevent-scroll" {...scrollableContentSwipePrevent}>
-                {categories.length > 0 &&
-                    <div ref={categoriesListRef} className="search__categories">
-                        {categories?.map(([category, categoryInfo]) =>
-                            <mi-chip
-                                icon={categoryInfo.iconUrl}
-                                background-color={primaryColor}
-                                content={categoryInfo.displayName}
-                                active={selectedCategory === category}
-                                onClick={() => categoryClicked(category)}
-                                key={category}>
-                            </mi-chip>
-                        )}
-                    </div>}
-                {showNotFoundMessage && <p className="search__error"> {t('Nothing was found')}</p>}
-                {searchResults.length > 0 &&
-                    <div className="search__results">
-                        {searchResults.map(location =>
-                            <ListItemLocation
-                                key={location.id}
-                                location={location}
-                                locationClicked={() => onLocationClicked(location)}
-                                isHovered={location?.id === hoveredLocation?.id}
-                            />
-                        )}
-                    </div>
-                }
-            </div>
+
+
+
+            { /* Horizontal list of Categories */}
+
+            {categories.length > 0 && <Categories onSetSize={onSetSize}
+                searchFieldRef={searchFieldRef}
+                getFilteredLocations={category => getFilteredLocations(category)}
+            />}
+
+
+
+
+            { /* Message shown if no search results were found */}
+
+            {showNotFoundMessage && <p className="search__error"> {t('Nothing was found')}</p>}
+
+
+
+            { /* Vertical list of search results. Scrollable. */}
+
+            {searchResults.length > 0 &&
+                <div className="search__results prevent-scroll" {...scrollableContentSwipePrevent}>
+                    {searchResults.map(location =>
+                        <ListItemLocation
+                            key={location.id}
+                            location={location}
+                            locationClicked={() => onLocationClicked(location)}
+                            isHovered={location?.id === hoveredLocation?.id}
+                        />
+                    )}
+                </div>
+            }
+
+
+
+            { /* Keyboard */}
+
             {isKeyboardVisible && isDesktop && <Keyboard ref={keyboardRef} searchInputElement={searchInput}></Keyboard>}
+
+
+
+            { /* Buttons to scroll in the list of search results if in kiosk context */}
+
+            {isOpen && isKioskContext && searchResults.length > 0 && createPortal(
+                <div className="search__scroll-buttons">
+                    <mi-scroll-buttons ref={scrollButtonsRef}></mi-scroll-buttons>
+                </div>,
+                document.querySelector('.mapsindoors-map')
+            )}
         </div>
     )
 }