+++ packages/map-template/src/components/VenueSelector/VenueSelector.jsx
@@ -9,6 +9,9 @@ import { ReactComponent as CloseIcon } from '../../assets/close.svg';
 import Venue from './Venue/Venue';
 import currentVenueNameState from '../../atoms/currentVenueNameState';
 import isLocationClickedState from '../../atoms/isLocationClickedState';
+import useSetCurrentVenueName from '../../hooks/useSetCurrentVenueName';
+import fitBoundsLocation from '../../helpers/fitBoundsLocation';
+import mapsIndoorsInstanceState from '../../atoms/mapsIndoorsInstanceState';
 
 /**
  * Show a list of Venues. The user can click on a Venue to select it.
@@ -22,8 +25,11 @@ function VenueSelector({ onOpen, onClose, active }) {
 
     const venueSelectorContentRef = useRef(null);
     const venues = useRecoilValue(venuesState);
+    const mapsIndoorsInstance = useRecoilValue(mapsIndoorsInstanceState);
 
-    const [currentVenueName, setCurrentVenueName] = useRecoilState(currentVenueNameState);
+    const setCurrentVenueName = useSetCurrentVenueName();
+
+    const currentVenueName = useRecoilValue(currentVenueNameState);
     const [, setIsLocationClicked] = useRecoilState(isLocationClickedState);
 
     /**
@@ -33,6 +39,7 @@ function VenueSelector({ onOpen, onClose, active }) {
      */
     const selectVenue = venue => {
         setCurrentVenueName(venue.name);
+        fitBoundsLocation(venue, mapsIndoorsInstance, 0, 0);
         toggle();
     };